Adding iter

+++ ObjDumper/ObjdTool.cs
@@ -8,36 +8,40 @@ namespace ObjDumper
     internal static class ObjdTool
     {
         public static void StartSh3(byte[] bytes, IDictionary<string, ParsedLine> res,
-            string dir, long s)
+            string dir, string id)
         {
-            var id = IterTool.GetId(s);
             if (res.ContainsKey(id)) return;
-            var name = Path.Combine(dir, $"sh3_{s:D8}.bin");
+            var name = Path.Combine(dir, $"sh3_{id}.bin");
             File.WriteAllBytes(name, bytes);
             var info = new ProcessStartInfo
             {
                 FileName = "sh-elf-objdump",
                 ArgumentList = { "-D", "-b", "binary", "-m", "sh3", "-z", name }
             };
+            if (!File.Exists(name))
+                return;
             var itm = StartAndGet(info);
-            res[id] = itm;
+            if (itm != null)
+                res[id] = itm;
             File.Delete(name);
         }
 
         public static void StartI86(byte[] bytes, IDictionary<string, ParsedLine> res,
-            string dir, long s)
+            string dir, string id)
         {
-            var id = IterTool.GetId(s);
             if (res.ContainsKey(id)) return;
-            var name = Path.Combine(dir, $"i86_{s:D8}.bin");
+            var name = Path.Combine(dir, $"i86_{id}.bin");
             File.WriteAllBytes(name, bytes);
             var info = new ProcessStartInfo
             {
                 FileName = "objdump",
                 ArgumentList = { "-D", "-Mintel,i8086", "-b", "binary", "-m", "i386", "-z", name },
             };
+            if (!File.Exists(name))
+                return;
             var itm = StartAndGet(info);
-            res[id] = itm;
+            if (itm != null)
+                res[id] = itm;
             File.Delete(name);
         }
     }
